🦺 Validações no CRUD para a entidade: Prospect
Implementadas as duas primeiras validações na Aplicação para a Entidade: "Prospect", já dentro dos Try/Catch Blocks desenvolvidos anteriormente e também com a padronização dos retornos relacionados para as devidas operações envolvidas. Para o 'Update' ser realizado é necessário que esse "Prospect" não tenha vinculados (através da FK - Foreign Key informada na criação do "Date" para ele) "Date"(s) que estejam com o 'Status == Requested'. Para o 'Delete' ser realizado é necessário que esse "Prospect" não tenha vinculado nenhum "Date", independente dos 'Status' (nesse caso é possível apenas a inativação dessa Entidade), no Back-End da Aplicação.

+++ backend/Domain/Requested/Handlers/DeleteProspectHandler.cs
@@ -21,14 +21,21 @@ using Domain.Requested.Repositories;
 
                         if(prospect != null) {
                             try {
-                                _prospectRepository.Delete(prospect);
+                                if(prospect.Dates.Count > 0) {
+                                    return new CommandResult<Unit>(
+                                        statusCode: HttpStatusCode.BadRequest,
+                                        statusHint: "BadRequest"
+                                    );
+                                } else {
+                                    _prospectRepository.Delete(prospect);
 
-                                    await _unityOfWork.Commit();
+                                        await _unityOfWork.Commit();
 
-                                        return new CommandResult<Unit>(
-                                            statusCode: HttpStatusCode.NoContent,
-                                            statusHint: "NoContent"
-                                        );
+                                            return new CommandResult<Unit>(
+                                                statusCode: HttpStatusCode.NoContent,
+                                                statusHint: "NoContent"
+                                            );
+                                }
                             } catch (Exception exception) {
                                 _unityOfWork.Rollback();
 
@@ -42,7 +49,7 @@ using Domain.Requested.Repositories;
                             return new CommandResult<Unit>(
                                 statusCode: HttpStatusCode.NotFound,
                                 statusHint: "NotFound"
-                            );       
+                            );
                 }
         }
     }
